src/requirements.txt
datastory == 0.1.14
db-dtypes == 1.1.1
google-cloud-bigquery == 3.11.3
google-cloud-secret-manager == 2.16.1
numpy == 1.25.0
pandas == 2.0.3
plotly == 5.15.0
scikit-learn == 1.3.0
